June 2017 - little fix to dev16-19 send value.
UPDATE: June 2017 - When I started converting my ha system from OpenHAB
to Home Assistant I noticed that this node had some unusual behavior
that I had to clean up. This was when I was first doing some basic MQTT
direct from the node to Home Assistant, not when I later introduced my
MySensors GW and code. Anyway, the issue was that when I used OpenHAB to
turn on the Elec Side Lights (or the other lights) OpenHAB would send
dev016 a 1, this node would turn on the lights correctly but respond
with "ON", not a matching 1.  OpenHAB was ok with that, but Home
Assistant wasn't. So I had to use GitHub Desktop on my ShedPC to get the
branch (deploy-to-shed-eth-node-21) of my code from last October etc
back onto my ShedPC, make a change to Eth_dosends.cpp such that it sent
back a 1 not an ON. To do that all I had to do was call sendInteger
rather than sendStatusInteger.
Now as this change was just a fixup, I have left it in the branch and
not merged it back into my MASTER code train.