made updates to ensure GUI kept in sync
Made some updates to RF_parseCmd.cpp so that if the node was perhaps in
STATIC_ONE_COLOUR mode and the user selected a STATIC_PATTERN in the HA
GUI, not only would the node change to that mode and display correct
pattern, but it would also update its copy of current291 and send it
back upstream so the MODE selector in the HA GUI would be updated
correctly.
I also made a minor change in setupLEDStrips() to correctly seperate the
setup code for remote vs local LED strips.