updated STATUS LED coding
Previously the BLUE STATUS LED was on for one sec then off for one sec.
I'd prefer it was on super breifly (e.g. few 100mSec) then off for a few
seconds as at the moment its on too much and to bright/distracting. So I
have modified the code that controls it and setup some new #defines that
control how often and how long it should be on.  In future I could
modify those on the fly perhaps to indicate an error via the Status LED.