RF msg fmt change & make work with SAMD RFM GW
I have changed the overall Node Code to use 32bit or long for most of
the parameters it puts in the struct Message, to send over RF, so it no
longer matters if running on AVR op SAMD, RF frame exact same.  This was
tested and works with my still in test SAMD RFM GW on SenseBender board.